edit data

+++ app/Controllers/Loan.php
@@ -284,7 +284,7 @@ class Loan extends BaseController
             $land_account_name = $this->SettingLandModel->getSettingLandByID($account_id);
             if ($charges_process != 0) {
                 $price = str_replace(',', '', $charges_process);
-                $land_account_cash_charges_process = $land_account_name->land_account_cash + intval($price);
+                $land_account_cash_charges_process = $land_account_name->land_account_cash + floatval($price);
 
                 $this->SettingLandModel->updateSettingLandByID($land_account_name->id, [
                     'land_account_cash' => $land_account_cash_charges_process,
@@ -309,7 +309,7 @@ class Loan extends BaseController
             $land_account_name = $this->SettingLandModel->getSettingLandByID($account_id);
             if ($charges_transfer != 0) {
                 $price = str_replace(',', '', $charges_transfer);
-                $land_account_cash_charges_transfer = $land_account_name->land_account_cash + intval($price);
+                $land_account_cash_charges_transfer = $land_account_name->land_account_cash + floatval($price);
 
                 $this->SettingLandModel->updateSettingLandByID($land_account_name->id, [
                     'land_account_cash' => $land_account_cash_charges_transfer,
@@ -334,7 +334,7 @@ class Loan extends BaseController
             $land_account_name = $this->SettingLandModel->getSettingLandByID($account_id);
             if ($charges_etc != 0) {
                 $price = str_replace(',', '', $charges_etc);
-                $land_account_cash_charges_etc = $land_account_name->land_account_cash + intval($price);
+                $land_account_cash_charges_etc = $land_account_name->land_account_cash + floatval($price);
 
                 $this->SettingLandModel->updateSettingLandByID($land_account_name->id, [
                     'land_account_cash' => $land_account_cash_charges_etc,
@@ -481,7 +481,7 @@ class Loan extends BaseController
         $charges_process_old = str_replace(',', '', $data->loan_payment_process);
         if ($charges_process_old != $charges_process) {
             if ($charges_process_old != 0) {
-                $land_account_cash_charges_process_old = $land_account_name_old->land_account_cash - intval($charges_process_old);
+                $land_account_cash_charges_process_old = $land_account_name_old->land_account_cash - floatval($charges_process_old);
 
                 $this->SettingLandModel->updateSettingLandByID($land_account_name_old->id, [
                     'land_account_cash' => $land_account_cash_charges_process_old,
@@ -503,7 +503,7 @@ class Loan extends BaseController
 
             $land_account_name = $this->SettingLandModel->getSettingLandByID($data->land_account_id);
             $price = str_replace(',', '', $charges_process);
-            $land_account_cash_charges_process = $land_account_name->land_account_cash + intval($price);
+            $land_account_cash_charges_process = $land_account_name->land_account_cash + floatval($price);
 
             $this->SettingLandModel->updateSettingLandByID($land_account_name->id, [
                 'land_account_cash' => $land_account_cash_charges_process,
@@ -528,7 +528,7 @@ class Loan extends BaseController
         if ($charges_transfer_old != $charges_transfer) {
             if ($charges_transfer_old != 0) {
                 $land_account_name = $this->SettingLandModel->getSettingLandByID($data->land_account_id);
-                $land_account_cash_charges_transfer_old = $land_account_name->land_account_cash - intval($charges_transfer_old);
+                $land_account_cash_charges_transfer_old = $land_account_name->land_account_cash - floatval($charges_transfer_old);
 
                 $this->SettingLandModel->updateSettingLandByID($land_account_name->id, [
                     'land_account_cash' => $land_account_cash_charges_transfer_old,
@@ -550,7 +550,7 @@ class Loan extends BaseController
 
             $land_account_name = $this->SettingLandModel->getSettingLandByID($data->land_account_id);
             $price = str_replace(',', '', $charges_transfer);
-            $land_account_cash_charges_transfer = $land_account_name->land_account_cash + intval($price);
+            $land_account_cash_charges_transfer = $land_account_name->land_account_cash + floatval($price);
 
             $this->SettingLandModel->updateSettingLandByID($land_account_name->id, [
                 'land_account_cash' => $land_account_cash_charges_transfer,
@@ -574,7 +574,7 @@ class Loan extends BaseController
         if ($loan_payment_other_old != $charges_etc) {
             if ($loan_payment_other_old != 0) {
                 $land_account_name = $this->SettingLandModel->getSettingLandByID($data->land_account_id);
-                $land_account_cash_loan_payment_other_old = $land_account_name->land_account_cash - intval($loan_payment_other_old);
+                $land_account_cash_loan_payment_other_old = $land_account_name->land_account_cash - floatval($loan_payment_other_old);
 
                 $this->SettingLandModel->updateSettingLandByID($land_account_name->id, [
                     'land_account_cash' => $land_account_cash_loan_payment_other_old,
@@ -596,7 +596,7 @@ class Loan extends BaseController
 
             $land_account_name = $this->SettingLandModel->getSettingLandByID($data->land_account_id);
             $price = str_replace(',', '', $charges_etc);
-            $land_account_cash_charges_etc = $land_account_name->land_account_cash + intval($price);
+            $land_account_cash_charges_etc = $land_account_name->land_account_cash + floatval($price);
 
             $this->SettingLandModel->updateSettingLandByID($land_account_name->id, [
                 'land_account_cash' => $land_account_cash_charges_etc,
@@ -699,7 +699,7 @@ class Loan extends BaseController
             $land_account_name = $this->SettingLandModel->getSettingLandByID($dataLoanByCode->land_account_id);
             $charges_process = str_replace(',', '', $dataLoanByCode->loan_payment_process);
 
-            $land_account_cash_charges_process = $land_account_name->land_account_cash - intval($charges_process);
+            $land_account_cash_charges_process = $land_account_name->land_account_cash - floatval($charges_process);
 
             $this->SettingLandModel->updateSettingLandByID($land_account_name->id, [
                 'land_account_cash' => $land_account_cash_charges_process,
@@ -723,7 +723,7 @@ class Loan extends BaseController
             $charges_transfer = str_replace(',', '', $dataLoanByCode->loan_tranfer);
 
             $land_account_name = $this->SettingLandModel->getSettingLandByID($dataLoanByCode->land_account_id);
-            $land_account_cash_charges_transfer = $land_account_name->land_account_cash - intval($charges_transfer);
+            $land_account_cash_charges_transfer = $land_account_name->land_account_cash - floatval($charges_transfer);
 
             $this->SettingLandModel->updateSettingLandByID($land_account_name->id, [
                 'land_account_cash' => $land_account_cash_charges_transfer,
@@ -747,7 +747,7 @@ class Loan extends BaseController
             $loan_payment_other = str_replace(',', '', $dataLoanByCode->loan_payment_other);
 
             $land_account_name = $this->SettingLandModel->getSettingLandByID($dataLoanByCode->land_account_id);
-            $land_account_cash_loan_payment_other = $land_account_name->land_account_cash - intval($loan_payment_other);
+            $land_account_cash_loan_payment_other = $land_account_name->land_account_cash - floatval($loan_payment_other);
 
             $this->SettingLandModel->updateSettingLandByID($land_account_name->id, [
                 'land_account_cash' => $land_account_cash_loan_payment_other,
